Beginning splitting
