+++ apps/ui/src/routes/coursePage.js
@@ -1,7 +1,7 @@
 import { Container, Paper } from '@mui/material';
 import { Grid } from '@mui/material';
 import { Navbar } from '../components/navbar.js';
-import { EntryRow, CourseList } from '../components/courses.js';
+import { CourseEntryRow, CourseList } from '../components/courses.js';
 import { APICoursesDelete, APICoursesGet, APICoursesPost } from '../api.js';
 
 import React from 'react';
@@ -60,7 +60,7 @@ export default class CoursePage extends React.Component {
                     <Grid container spacing={2} padding={1}>
                         <Grid item xs={12}>
                             <Paper sx={{marginBottom: '8px'}} elevation={1}>
-                                <EntryRow
+                                <CourseEntryRow
                                     newCourseCallback={this.newCourse}
                                     userid={this.state.userid}
                                 />
